requirements.txt
fastapi
uvicorn
frosty-ai==1.5
python-dotenv
